refactor: menugroup 패키지 생성
